requirements.txt
Flask==2.0.1
Pillow==8.2.0
scikit-image==0.18.1
dlib==19.22.0
scikit-learn==0.24.2
opencv-python-headless==4.5.2.52
numpy==1.20.3
requests==2.25.1
